replace load_case with re_interpolate_solution
